specs - fix after :all callback
